+++ cmd/copy/workspaces-states.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package copy
 
 import (
@@ -176,9 +179,9 @@ func copyStates(c tfclient.ClientContexts, NumberOfStates int) error {
 	}
 
 	// Get/Check if Workspace map exists
-	wsMapCfg, err := helper.ViperStringSliceMap("workspace-map")
+	wsMapCfg, err := helper.ViperStringSliceMap("workspaces-map")
 	if err != nil {
-		fmt.Println("invalid input for workspace-map")
+		fmt.Println("invalid input for workspaces-map")
 	}
 
 	// Get the destination target workspaces
@@ -252,8 +255,8 @@ func copyStates(c tfclient.ClientContexts, NumberOfStates int) error {
 					lockWorkspace(tfclient.GetClientContexts(), destWorkspaceId)
 					fmt.Printf("Migrating state version %v serial %v for workspace Src: %v Dst: %v\n", srcstate.StateVersion, newSerial, srcworkspace.Name, destWorkSpaceName)
 					srcstate, err := c.DestinationClient.StateVersions.Create(c.DestinationContext, destWorkspaceId, tfe.StateVersionCreateOptions{
-						Type:             "",
-						Lineage:          new(string),
+						Type: "",
+						//Lineage:        Optional attribute, must be left empty or match source Lineage
 						MD5:              tfe.String(md5String),
 						Serial:           &newSerial,
 						State:            tfe.String(stringState),