+++ cmd/copy/workspaces-vcs.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package copy
 
 import (
@@ -42,9 +45,9 @@ func createVCSConfiguration(c tfclient.ClientContexts, vcsConfig map[string]stri
 		}
 
 		// Get/Check if Workspace map exists
-		wsMapCfg, err := helper.ViperStringSliceMap("workspace-map")
+		wsMapCfg, err := helper.ViperStringSliceMap("workspaces-map")
 		if err != nil {
-			fmt.Println("Invalid input for workspace-map")
+			fmt.Println("Invalid input for workspaces-map")
 		}
 
 		// For each source workspace with a VCS connection, compare the source oauth ID to the