+++ cmd/list/list.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package list
 
 import (
@@ -6,7 +9,8 @@ import (
 
 var (
 	side    string
-	
+	jsonOut bool
+
 	ListCmd = &cobra.Command{
 		Use:   "list",
 		Short: "List command",
@@ -17,4 +21,5 @@ var (
 func init() {
 
 	ListCmd.PersistentFlags().StringVar(&side, "side", "", "Specify source or destination side to process")
+	ListCmd.PersistentFlags().BoolVar(&jsonOut, "json", false, "Print the output in JSON format")
 }